requirements.txt
astroid==2.1.0
backports.csv==1.0.7
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
django-bootstrap3==11.0.0
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
django-filter==2.0.0
django-import-export==1.2.0
django-jquery==3.1.0
django-model-utils==3.1.2
django-multiselectfield==0.1.8
django-quiz-app==0.5.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.11
et-xmlfile==1.0.1
filelock==3.0.10
idna==2.7
isort==4.3.4
jdcal==1.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
odfpy==1.4.0
openpyxl==2.5.12
Pillow==5.3.0
pluggy==0.8.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
py==1.7.0
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
pytz==2018.7
PyYAML==5.1.2
requests==2.20.1
securionpay==1.0.1
six==1.11.0
sqlparse==0.3.0
tablib==0.13.0
toml==0.10.0
tox==3.5.3
urllib3==1.24.1
virtualenv==16.1.0
wrapt==1.10.11
xlrd==1.2.0
xlwt==1.3.0
